requirements.txt
# Core ML framework
torch>=2.0.0
torchvision>=0.15.0

# Numerical computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
Pillow>=9.0.0

# FLIR thermal imaging
flirpy>=0.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.60.0

# Configuration
pyyaml>=6.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Monitoring (optional)
tensorboard>=2.8.0
